refactor: centralize exit code mapping
