+++ crates/cli/src/lib.rs
@@ -75,6 +75,27 @@ pub fn exit_code_from_error_kind(kind: clap::error::ErrorKind) -> ExitCode {
     }
 }
 
+pub fn exit_code_from_engine_error(e: &EngineError) -> ExitCode {
+    use std::io::ErrorKind;
+    match e {
+        EngineError::Io(err) => match err.kind() {
+            ErrorKind::TimedOut | ErrorKind::WouldBlock => ExitCode::ConnTimeout,
+            ErrorKind::ConnectionRefused
+            | ErrorKind::AddrNotAvailable
+            | ErrorKind::NetworkUnreachable
+            | ErrorKind::ConnectionAborted
+            | ErrorKind::ConnectionReset
+            | ErrorKind::NotConnected
+            | ErrorKind::HostUnreachable
+            | ErrorKind::NetworkDown => ExitCode::SocketIo,
+            _ => ExitCode::Protocol,
+        },
+        EngineError::MaxAlloc => ExitCode::Malloc,
+        EngineError::Exit(code, _) => *code,
+        _ => ExitCode::Protocol,
+    }
+}
+
 pub fn handle_clap_error(cmd: &clap::Command, e: clap::Error) -> ! {
     use clap::error::ErrorKind;
     let kind = e.kind();
@@ -1814,6 +1835,63 @@ mod tests {
     fn exit_code_handles_unknown_error_kind() {
         let kind = clap::error::ErrorKind::DisplayHelpOnMissingArgumentOrSubcommand;
         assert_eq!(exit_code_from_error_kind(kind), ExitCode::SyntaxOrUsage);
+    }
+
+    #[test]
+    fn maps_error_kinds_to_exit_codes() {
+        use clap::error::ErrorKind::*;
+        let cases = [
+            (InvalidValue, ExitCode::Unsupported),
+            (UnknownArgument, ExitCode::SyntaxOrUsage),
+            (InvalidSubcommand, ExitCode::SyntaxOrUsage),
+            (NoEquals, ExitCode::SyntaxOrUsage),
+            (ValueValidation, ExitCode::SyntaxOrUsage),
+            (TooManyValues, ExitCode::SyntaxOrUsage),
+            (TooFewValues, ExitCode::SyntaxOrUsage),
+            (WrongNumberOfValues, ExitCode::SyntaxOrUsage),
+            (ArgumentConflict, ExitCode::SyntaxOrUsage),
+            (MissingRequiredArgument, ExitCode::SyntaxOrUsage),
+            (MissingSubcommand, ExitCode::SyntaxOrUsage),
+            (InvalidUtf8, ExitCode::SyntaxOrUsage),
+            (
+                DisplayHelpOnMissingArgumentOrSubcommand,
+                ExitCode::SyntaxOrUsage,
+            ),
+            (DisplayHelp, ExitCode::Ok),
+            (DisplayVersion, ExitCode::Ok),
+            (Io, ExitCode::FileIo),
+            (Format, ExitCode::FileIo),
+        ];
+
+        for (kind, expected) in cases {
+            assert_eq!(exit_code_from_error_kind(kind), expected);
+        }
+    }
+
+    #[test]
+    fn transient_network_errors_map_to_conn_timeout() {
+        use std::io::ErrorKind;
+        let kinds = [
+            ErrorKind::TimedOut,
+            ErrorKind::ConnectionRefused,
+            ErrorKind::AddrNotAvailable,
+            ErrorKind::NetworkUnreachable,
+            ErrorKind::WouldBlock,
+            ErrorKind::ConnectionAborted,
+            ErrorKind::ConnectionReset,
+            ErrorKind::NotConnected,
+            ErrorKind::HostUnreachable,
+            ErrorKind::NetworkDown,
+        ];
+
+        for kind in kinds {
+            let err = EngineError::Io(std::io::Error::from(kind));
+            assert_eq!(
+                exit_code_from_engine_error(&err),
+                ExitCode::ConnTimeout,
+                "{kind:?} did not map to ConnTimeout",
+            );
+        }
     }
 
     #[test]
